rename files fix double definition for clamp util
