+++ lib/utils/utils.hpp
@@ -1,0 +1,6 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+int clamp(int value, int minValue, int maxValue, bool wrap = false);
+
+#endif // UTILS_H
